Included ability to choose matrix sizes.
Need to implement this for the rest of the code, and allow for the functions to take the largest value for the row/ column, and derive whether or not it's possible for the operation to be done. For example, the columns of the 1st must equal the rows of the 2nd for matrix operations to be possible.